+++ noodles-sam/CHANGELOG.md
@@ -28,7 +28,8 @@
   * sam/header: The key for `ReferenceSequences` is now a
     `reference_sequence::Name`.
 
-  * sam/header/record: Add record IDs to parse errors.
+  * sam/header/record: Add record IDs to `ParseError::InvalidReferenceSequence`
+    and `ParseError::InvalidReadGroup`.
 
   * sam/header/record/value/map/reference_sequence: `ReferenceSequence::new`
     requires a nonzero length.